Fixed DCT
It was never broken in the first place! It just turned out that the new
Python function for measuring how well it perform was making it look
worse than it actually is. By using my old Matlab function and comparing
at the binary level, it is evident that the DCT steganography is
actually quite successful.