Set control to use uint8 instead of double
This is because I noticed that the LSB steganography technique
consistently gave an even higher PSNR than the control, which does not
make sense at all. I then remembered that LSB steganography uses the
uint8 data type, whereas all the others have to use double because of
the transformations. By changing the control to use uint8 instead of
double for the image values, it maintains a higher PSNR, which is
correct.